Create data health module

Inspection of the quality of the records and demonstrations of their enrichment

Usage

create_data_health_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  PlantClassification_filepath = NULL,
  coordinates = NA,
  output_dir = NULL,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL
)

Arguments

enriched_report

The enriched report of a living collection (created in BGsmartR)

collection

The name of the living collection

wgsrpd3

World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from rWCVP package.

PlantClassification_filepath

filepath to family classification. If NULL we use the default contained within BGSmartR.

coordinates

The coordinates of the collection (long, lat).

output_dir

The output directory

native

string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.

extinct

Flag (TRUE/FALSE) for whether to include extinct geographic locations.

doubtful_locations

Flag (TRUE/FALSE) for whether to include doubtful locations.

earliest_allowable_record

Number, corresponding to the minimum allowable year in the "AccYear" column.

old_accession_year_codes

Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on earliest_allowable_record.