Create rarity and duplication module

Degree of rarity of a plant in cultivation globally and duplication of individuals per species in a collection.

Usage

create_duplication_report(
  enriched_report,
  collection = NULL,
  min_year = 1970,
  join_bigger_than = 4,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  output_dir = NULL
)

Arguments

enriched_report

The enriched report of a living collection (created in BGsmartR)

collection

The name of the living collection

min_year

The minimum year for analyses over time.

join_bigger_than

When exploring accessions per item, etc the value at which we merge larger values for simplified pie charts. Default is 4.

native

string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.

extinct

Flag (TRUE/FALSE) for whether to include extinct geographic locations.

doubtful_locations

Flag (TRUE/FALSE) for whether to include doubtful locations.

output_dir

The output directory