create_last_update_report()

This function can be used to create the last update report of a living collection. The report informs on the items that have missing/placeholder/future item status date and creates figures of the when items were last updated.

Usage

create_last_update_report(
  enriched_report,
  collection = NULL,
  report_kind = "static",
  output_file = NULL,
  output_dir = NULL,
  export_data = FALSE,
  separate_figure_folder = TRUE,
  table_font_size = 10,
  ggtheme = NULL,
  value_on_fig = FALSE,
  save_excel = FALSE,
  scale_colour_continuous = ggplot2::scale_colour_viridis_c,
  scale_colour_discrete = ggplot2::scale_colour_viridis_d,
  scale_colour_binned = ggplot2::scale_colour_viridis_b,
  scale_fill_continuous = ggplot2::scale_fill_viridis_c,
  scale_fill_discrete = ggplot2::scale_fill_viridis_d,
  scale_fill_binned = ggplot2::scale_fill_viridis_b,
  reference_docx = NULL
)

Arguments

enriched_report

The enriched (via BGSmartR) report of a LC.

collection

The name of the LC

report_kind

The find of report to create, either static or interactive.

output_file

The file path of the native report.

output_dir

The output directory

export_data

Flag (TRUE/FALSE) for whether to create a .rda and .xlsx files containing the data used to create the figures in the plot.

separate_figure_folder

Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.

table_font_size

Font size for tables.

ggtheme

gg plot theme to be applied to all ggplots, see ggthemes package for pre-set themes.

value_on_fig

Flag (TRUE/FALSE) for whether to include values of static plots.

save_excel

Flag (TRUE/FALSE) if export_data == TRUE, save excel allows for a 'xlsx' version of the data to be created or not. Note that for particualarly large collections, saving an excel version may cause the function to fail.

scale_colour_continuous, scale_colour_discrete, scale_colour_binned, scale_fill_continuous, scale_fill_discrete, scale_fill_binned

scales for ggplot. Default is to use viridis.

reference_docx

path to a .docx file whose style (design) the report copies.