Enrich a collection using enrichment database

Enrich a collection using enrichment database

Usage

enrich_collection_from_enrich_database(
  collection,
  enrich_database,
  ...,
  columns_to_enrich = NA,
  add_to_column_name = "Enrich_",
  add_match_details = TRUE
)

Arguments

collection

A data frame containing a collection.

enrich_database

A data frame of enriching information.

...

Arguments (i.e., attributes) used in the matching algorithm (passed along to nested functions). See match_collection_to_enrich_database().

columns_to_enrich

Column names of enrich_database that are enriched to collection.

add_to_column_name

A string to prepend to new columns in collection created by enrichment.

add_match_details

A flag (TRUE/FALSE) for whether the match details also want to be added to the enriched collection.

Value

collection enriched with new columns after matching to enrich_database

Details

This function takes a collection and "enriches" it by adding new information taken from enrich_database by matching via taxonomic names. The matching is performed by match_collection_to_enrich_database(), and inputs can be passed to this function via ....

The enrichment columns to add can be specified by columns_to_enrich, and you can include th matching details via add_match_details. To make sure the enrichment column names don't clash with names already in the collection each is prepended with add_to_column_name.