Find best author matches

Find best author matches

Usage

match_authors(
  collection_author,
  enriched_database_authors,
  partial_method = "most words",
  ...
)

Arguments

collection_author

The author from the collection wanted to be matched

enriched_database_authors

Author options from the enrich_database.

partial_method

Either 'most words or 'any words', defining the method used to find partial matches.

...

Arguments (i.e., attributes) used in the matching algorithm (passed along to nested fuctions).

Value

a list of length 2 with:

$wanted is a logical (TRUE/FALSE) vector with length length(enriched_database_authors) corresponding to the enriched database authors that most match the collection author.

$message detailing whether the author match was exact, partial or no match was found.

Examples

collection_author = 'Schult'
enriched_database_authors = c("(Lour.) Schult", "Borhidi & E.Martinez")
match_authors(collection_author, enriched_database_authors)
#> $wanted
#> [1]  TRUE FALSE
#> 
#> $message
#> [1] "(Partial author <most words>)"
#>