Match collection to IUCN Red List of Threatened Species via taxonomic name

Match collection to IUCN Red List of Threatened Species via taxonomic name

Usage

match_collection_to_iucnRedlist(
  collection,
  iucnRedlist,
  taxon_name_column = "TaxonName",
  taxon_name_full_column = NA,
  taxon_author_column = NA,
  typo_method = "fast",
  try_add_split = TRUE,
  try_fix_hybrid = TRUE,
  try_rm_autonym = TRUE,
  ...,
  enrich_taxon_name_column = "scientific_name",
  enrich_display_in_message_column = "taxonid",
  enrich_plant_identifier_column = "taxonid"
)

Arguments

collection

A data frame containing a collection.

iucnRedlist

IUCN Red List of Threatened Species database, obtained using the function XXXX.

taxon_name_column

The name of the column in the collection corresponding to taxonomic names.

taxon_name_full_column

The name of the column in the collection corresponding to joined taxonomic names and authors.

taxon_author_column

The name of the column in the collection corresponding to the authors of the taxonomic names.

typo_method

Flag for whether we search for typos.

try_add_split

Flag (TRUE/FALSE) for whether we search for missing f./var./subsp.

try_fix_hybrid

Flag (TRUE/FALSE) for whether we search for hybrid issues.

try_rm_autonym

Flag (TRUE/FALSE) for whether we try removing autonyms.

...

Arguments (i.e., attributes) used in the matching algorithm (passed along to nested functions). Examples include enrich_taxon_authors_column, enrich_display_in_message_column and enrich_plant_identifier_column.

enrich_taxon_name_column

The name of the column in the iucnRedlist corresponding to taxonomic names.Default value is scientific_name.

enrich_display_in_message_column

The name of the column in iucnRedlist that contains values to show in the matching messages. Default value is taxonid.

enrich_plant_identifier_column

The name of the column in iucnRedlist that corresponds to record identifier. Default value is taxonid.

Value

A list of length seven containing:

$match the index of the record in iucnRedlist which matches the record in the collection database.

$details_short a simplified message detailing the match.

$match_taxon_name a longer format message detailing the match.

$original_authors The author/s (extracted) from the collection database.

$match_authors The author/s of the matched record in iucnRedlist.

$author_check Either Identical, Partial or Different (No Match if a match to iucnRedlist cannot be found). A message informing the similarity of the collection's taxon authors and the authors found in iucnRedlist. Author similarity is found using the function author_check().

match = taxon_match_full,

Details

This function allows matching of a collection's database to IUCN Red List of Threatened Species database. This function relies of matching functions found that are documented in match_single(), and is broadly similar to the function match_collection_to_wcvp().

Note that by default the matching functions use column names from (WCVP) therefore these will often require changing prior to matching unless you change the column names in iucnRedlist to concure with wcvp column names.

Within the algorithm there exists methods to improve the matching such as trying to change infraspecific levels (e.g var. to subsp.) or adding hybridisation. These methods can be turned on and off using try_add_split, try_fix_hybrid, try_rm_autonym and typo_method.