Matching functions

Functions used to match taxonomic names from a collection to exterior databases (POWO's WCVP, IUCN Redlist)

Usage

shorten_message(messages)

known_not_in_wcvp(taxon_names)

match_single_wcvp(taxon_names, wcvp, wcvp_search_index)

match_taxon_status(author_match, corres_POWO, current_message = "")

get_match_from_multiple(taxon_name_and_author, wcvp_mult)

match_mult_wcvp(
  taxon_names,
  taxon_authors,
  wcvp,
  wcvp_search_index,
  show_progress = TRUE
)

match_rm_autonym(
  taxon_names,
  taxon_authors,
  wcvp,
  single_indices = NA,
  mult_indices = NA
)

match_error(
  new_taxon_names,
  match_details,
  original_author,
  wcvp,
  current_message = ""
)

add_splitter(taxon_names, taxon_authors, wcvp)

match_hybrid_issue(taxon_names, taxon_authors, wcvp)

convert_to_accepted_name(original_match, wcvp)

check_taxon_typo(taxon, wcvp = NA, typo_df = BGSmartR::typo_list, fast = T)

match_typos(
  taxon_names,
  taxon_authors,
  wcvp,
  single_indices = NA,
  mult_indices = NA,
  typo_method = "fast"
)

match_all_issue(
  taxon_names,
  taxon_authors,
  do_taxon_author,
  wcvp,
  single_indices = NA,
  mult_indices = NA,
  try_add_split = TRUE,
  try_fix_hybrid = TRUE,
  try_rm_autonym = TRUE
)

Arguments

messages

messages about matching

taxon_names

Vector of taxonomic names.

wcvp

cleaned POWO database

wcvp_search_index

Indices of wcvp that we want to search for a match

author_match

vector of true of false whether we have a match

corres_POWO

the corresponding entries in POWO (rows must equal length author match)

current_message

the current message

taxon_name_and_author

the pair of taxon name and taxon name full.

wcvp_mult

POWO database restricted to records that do not have a unique taxon name.

taxon_authors

a vector of full taxon names (corresponding to taxon_names)

show_progress

Flag for whether we show progress bar.

single_indices

wcvp indices with a unique taxon name

mult_indices

wcvp indices with non-unique taxon name

new_taxon_names

new_taxon_names

match_details

match_details

original_author

original_author

original_match

a vector of indices corresponding to a match between a gardens database and POWO.

taxon

The taxon_name

typo_df

A data frame with two columns of typos and their fixes.

fast

A flag for whether we want to search only the typo_df (TRUE) or also search directly (FALSE)

typo_method

method for finding typos

do_taxon_author

do_taxon_author

try_add_split

Flag for whether we search for missing f./var./subsp.

try_fix_hybrid

Flag for whether we search for hybrid issues.

try_rm_autonym

Flag for whether we try removing autonyms.