Create data health module

Inspection of the quality of the records and demonstrations of their enrichment

Usage

create_data_health_report(
  enriched_report,
  collection = NULL,
  wgsrpd3 = NULL,
  PlantClassification_filepath = NULL,
  report_kind = "static",
  coordinates = NA,
  output_file = NULL,
  output_dir = NULL,
  export_data = FALSE,
  separate_figure_folder = TRUE,
  native = "Naturally occurring only",
  extinct = TRUE,
  doubtful_locations = FALSE,
  table_font_size = 10,
  ggtheme = NULL,
  value_on_fig = FALSE,
  earliest_allowable_record = 1700,
  old_accession_year_codes = NULL,
  scale_colour_continuous = ggplot2::scale_colour_viridis_c,
  scale_colour_discrete = ggplot2::scale_colour_viridis_d,
  scale_colour_binned = ggplot2::scale_colour_viridis_b,
  scale_fill_continuous = ggplot2::scale_fill_viridis_c,
  scale_fill_discrete = ggplot2::scale_fill_viridis_d,
  scale_fill_binned = ggplot2::scale_fill_viridis_b,
  reference_docx = NULL
)

Arguments

enriched_report

The enriched report of a living collection (created in BGsmartR)

collection

The name of the living collection

wgsrpd3

World Geographical Scheme for Recording Plant Distributions. Containing polygons of each level 3 area. Can be obtained from rWCVP package.

PlantClassification_filepath

filepath to family classification. If NULL we use the default contained within BGSmartR.

report_kind

The find of report to create, either static or interactive.

coordinates

The coordinates of the collection (long, lat).

output_file

The file path of the native report.

output_dir

The output directory

export_data

Flag (TRUE/FALSE) for whether to create a .rda and .xlsx files containing the data used to create the figures in the plot.

separate_figure_folder

Flag (TRUE/FALSE) for whether the figures should also be outputted in seperate files.

native

string where 'Naturally occurring only' or 'Introduced only' reduces the locations used to determine native plants.

extinct

Flag (TRUE/FALSE) for whether to include extinct geographic locations.

doubtful_locations

Flag (TRUE/FALSE) for whether to include doubtful locations.

table_font_size

Font size for tables.

ggtheme

gg plot theme to be applied to all ggplots, see ggthemes package for pre-set themes.

value_on_fig

Flag (TRUE/FALSE) for whether to include values of static plots.

earliest_allowable_record

Number, corresponding to the minimum allowable year in the "AccYear" column.

old_accession_year_codes

Vector of numbers corresponding to values within "AccYear" that indicate that a record is old with unknown accession year. For our analyses over time we assume these records are accessioned on earliest_allowable_record.

scale_colour_continuous, scale_colour_discrete, scale_colour_binned, scale_fill_continuous, scale_fill_discrete, scale_fill_binned

scales for ggplot. Default is to use viridis.

reference_docx

path to a .docx file whose style (design) the report copies.