Match collection to World Checklist of Vascular Plants via taxonomic name

Match collection to World Checklist of Vascular Plants via taxonomic name

Usage

match_collection_to_wcvp(
  collection,
  wcvp,
  taxon_name_column = "TaxonName",
  taxon_name_full_column = NA,
  taxon_author_column = NA,
  enrich_taxon_authors_column = "taxon_authors_simp",
  typo_method = "All",
  do_add_split = TRUE,
  do_fix_hybrid = TRUE,
  do_rm_autonym = TRUE,
  do_convert_accepted = TRUE,
  matching_criterion = BGSmartR::additional_wcvp_matching,
  ...
)

Arguments

collection

A data frame containing a collection.

wcvp

World Checklist of Vascular Plants (WCVP) database, obtained using the function import_wcvp_names().

taxon_name_column

The name of the column in the collection corresponding to taxonomic names.

taxon_name_full_column

The name of the column in the collection corresponding to joined taxonomic names and authors.

taxon_author_column

The name of the column in the collection corresponding to the authors of the taxonomic names.

enrich_taxon_authors_column

The name of the column in enrich_database that corresponds to the authors of taxonomic names. Default value is taxon_authors_simp.

typo_method

Either 'All', 'Data frame only','Data frame + common', detailing the level of typo finding required.

do_add_split

Flag (TRUE/FALSE) for whether we search for missing f./var./subsp.

do_fix_hybrid

Flag (TRUE/FALSE) for whether we search for hybrid issues.

do_rm_autonym

Flag (TRUE/FALSE) for whether we try removing autonyms.

do_convert_accepted

Flag for whether we convert to accepted names in wcvp

matching_criterion

A function used to chose the best method from extracts of the wcvp$wcvp_names.

...

Arguments (i.e., attributes) used in the matching algorithm (passed along to nested fuctions). Examples include, enrich_display_in_message_column and enrich_plant_identifier_column.

Value

A list of length seven containing:

$match the index of the record in wcvp$wcvp_names which matches the record in the collection database.

$details_short a simplified message detailing the match.

$match_taxon_name a longer format message detailing the match.

$original_authors The author/s (extracted) from the collection database.

$match_authors The author/s of the matched record in wcvp$wcvp_names.

$author_check Either Identical, Partial or Different (No Match if a match to wcvp cannot be found). A message informing the similarity of the collection's taxon authors and the authors found in wcvp$wcvp_names. Author similarity is found using the function author_check().

Details

This function allows matching of a collection's database to World Checklist of Vascular Plants (WCVP) database. For details of how the matching algorithm works see Method of Matching taxonomic records vignette (Matching.Rmd).